chore(corel): make documentDiff component agnostic to releases
